requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
assertpy==1.1
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
exceptiongroup==1.1.1
Faker==18.11.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
lxml==4.9.2
numpy==1.25.0
outcome==1.2.0
packaging==23.1
pandas==2.0.3
pluggy==1.2.0
PyPika==0.48.9
PySocks==1.7.1
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
tzdata==2023.3
urllib3==2.0.3
webdriver-manager==3.8.6
wsproto==1.2.0
